Fix: Move test assets to Fixtures namespace
