+++ tests/Fixtures/FooSignedRequestHelper.php
@@ -21,18 +21,14 @@
  * DEALINGS IN THE SOFTWARE.
  *
  */
-namespace Facebook\Tests;
+namespace Facebook\Tests\Fixtures;
 
-use Facebook\Http\GraphRawResponse;
-use Facebook\HttpClients\FacebookHttpClientInterface;
+use Facebook\Helpers\FacebookSignedRequestFromInputHelper;
 
-class MyFooBatchClientHandler implements FacebookHttpClientInterface
+class FooSignedRequestHelper extends FacebookSignedRequestFromInputHelper
 {
-    public function send($url, $method, $body, array $headers, $timeOut)
+    public function getRawSignedRequest()
     {
-        return new GraphRawResponse(
-            "HTTP/1.1 200 OK\r\nDate: Mon, 19 May 2014 18:37:17 GMT",
-            '[{"code":"123","body":"Foo"},{"code":"1337","body":"Bar"}]'
-        );
+        return null;
     }
 }
